chore(config): migrate remote agent client configuration
Build remote Agent IPC clients from typed command port, message size,
and vsock settings while authentication paths remain on the
compatibility map.

Move vsock validation to the binary and share the resolved client
configuration across bootstrap, environment providers, and host tags.
Mark the migrated inventory keys as typed configuration consumers,
including `expected_tags_duration`, which host tags now reads from the
typed model rather than the raw map.

Out-of-range `cmd_port` and negative `agent_ipc.grpc_max_message_size`
now fail the startup translation gate instead of being narrowed to fit.
This matches how the other migrated ports already behave.

Drop `connect_retry_attempts` and `connect_retry_backoff`. Neither key
was in the Agent schema or our inventory, so the Core Agent never sent
them. Their former defaults, ten attempts and two seconds, are now
constants next to the one call site that retries.

Rename the client constructor to `RemoteAgentClient::connect`, and
rename the wrappers that take a typed client configuration to `new`.
Only the builders that still read the raw map keep the
`from_configuration` name. Name the resolved CID field `vsock_cid` so
it no longer collides with the `vsock_addr()` accessor.

+++ lib/agent-data-plane-config/src/control.rs
@@ -109,17 +109,27 @@ pub struct Logging {
     pub file_max_size: u64,
 }
 
-/// Bootstrap IPC and remote-agent connection parameters, read before runtime authority exists.
+/// IPC and remote-agent connection parameters, read once at bootstrap before runtime authority
+/// exists and again from the authoritative configuration once it does.
+///
+/// The derived `Default` is all zeroes and serves only as the starting point for translation. The
+/// effective default of each field is the one the Datadog schema declares, noted per field below.
 #[derive(Clone, Debug, Default, PartialEq, Serialize)]
 pub struct ControlIpc {
     /// TCP port the command API listens on.
+    ///
+    /// Defaults to `5001`.
     pub cmd_port: u16,
 
     /// vsock address used for guest/host IPC.
+    ///
+    /// Defaults to empty, which reaches the Core Agent over TCP on localhost at `cmd_port`.
     pub vsock_addr: String,
 
     /// Maximum gRPC message size, in bytes, accepted over the remote-agent IPC channel.
-    pub grpc_max_message_size: i64,
+    ///
+    /// Defaults to `134217728` (128 MiB).
+    pub grpc_max_message_size: usize,
 
     /// Timeout for establishing a connection to the container runtime interface.
     pub cri_connection_timeout: i64,
